+++ app/Plugin/Configuration_Base.php
@@ -1,0 +1,57 @@
+<?php
+/**
+ * File as base for each mode.
+ *
+ * @package external-files-in-media-library
+ */
+
+namespace ExternalFilesInMediaLibrary\Plugin;
+
+// prevent direct access.
+defined( 'ABSPATH' ) || exit;
+
+/**
+ * Define the base object for configurations.
+ */
+class Configuration_Base {
+	/**
+	 * Name of this object.
+	 *
+	 * @var string
+	 */
+	protected string $name = '';
+
+	/**
+	 * Return the name of this object.
+	 *
+	 * @return string
+	 */
+	public function get_name(): string {
+		return $this->name;
+	}
+
+	/**
+	 * Return the title of this object.
+	 *
+	 * @return string
+	 */
+	public function get_title(): string {
+		return '';
+	}
+
+	/**
+	 * Return additional hints for the dialog to set this mode.
+	 *
+	 * @return array<int,string>
+	 */
+	public function get_dialog_hints(): array {
+		return array();
+	}
+
+	/**
+	 * Save the configuration this mode defines.
+	 *
+	 * @return void
+	 */
+	public function run(): void {}
+}